converted decimal? to int? and SQL col

+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.UnitTests/Services/EditValidation/EmploymentPriceValidationTests.cs
@@ -35,7 +35,7 @@ namespace SFA.DAS.CommitmentsV2.UnitTests.Services.EditValidation
             var employEndDate = fixture.Apprenticeship.EndDate.Value.AddMonths(-1);
             var request = fixture.CreateValidationRequest(deliveryModel: DeliveryModel.PortableFlexiJob,
                 employmentEndMonth: employEndDate.Month, employmentEndYear: employEndDate.Year);
-            request.EmploymentPrice = fixture.Apprenticeship.Cost + 1;
+            request.EmploymentPrice = (int)fixture.Apprenticeship.Cost + 1;
 
             var result = await fixture.Validate(request);
