+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2/Domain/Entities/EditApprenticeshipValidation/EditApprenticeshipValidationRequest.cs
@@ -23,6 +23,6 @@ namespace SFA.DAS.CommitmentsV2.Domain.Entities.EditApprenticeshipValidation
         public string ProviderReference { get; set; }
         public string Email { get; set; }
         public DateTime? EmploymentEndDate { get; set; }
-        public decimal? EmploymentPrice { get; set; }
+        public int? EmploymentPrice { get; set; }
     }
 }
