+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2/Models/ApprenticeshipUpdate.cs
@@ -17,7 +17,7 @@ namespace SFA.DAS.CommitmentsV2.Models
         public string Email { get; set; }
         public DeliveryModel? DeliveryModel { get; set; }
         public DateTime? EmploymentEndDate { get; set; }
-        public decimal? EmploymentPrice { get; set; }
+        public int? EmploymentPrice { get; set; }
         public ProgrammeType? TrainingType { get; set; }
         public string TrainingCode { get; set; }
         public string TrainingCourseVersion { get; set; }
